requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
get==2019.4.13
gunicorn==19.9.0
idna==2.8
post==2019.4.13
public==2019.4.13
pycodestyle==2.5.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
